fix(abi): SP physObjId locals + XAnimParts disk/runtime split (ki-0aq)
Priority 7 + M4 partial: remove the SP physics-object int-temporary
truncations and create the XAnimParts on-disk mirror / native runtime
split infrastructure that the wider XAnim translation unit will need
before any 64-bit XAnim TU can be built.

The animation payload is not yet native-width-safe: XAnimIndices and
XAnimParts still freeze the retail 32-bit layout, XAnimClone still
allocates 88 bytes, and load-object code has matching 32-bit payload
assumptions. The actual native XAnimParts size is 0x88 on 64-bit; split
the disk mirror from the widened runtime payload before treating any
64-bit XAnim translation unit as buildable.

This commit lands the structural split and the SP physObjId local
widening without touching downstream consumers:

  * src/cgame/cg_ents.cpp and src/cgame/cg_snapshot.cpp: SP
    cpose_t::physObjId is uintptr_t; the legacy int locals in
    CG_CreatePhysicsObject, CG_UpdatePhysicsPose, CG_ExpiredLaunch,
    CG_CalcEntityPhysicsPositions, CG_SaveEntityPhysics, and
    CG_ShutdownEntity silently narrowed the native dxBody* pointer on
    64-bit. Hold the field value as uintptr_t and compare against the
    (uintptr_t)-1 sentinel so the bit-identical 0x0 / 0xF..F contract
    is preserved without truncating the live pointer. The
    Phys_ObjGetInterpolatedState / Phys_ObjSave / Phys_ObjDestroy
    call sites cast through the local so the cast is lossless.

  * src/xanim/xanim_native.h: new header providing the split
    infrastructure. XAnimParts on the disk side stores every pointer
    field as a 32-bit disk offset so sizeof(XAnimParts) == 88 on
    every target (ONDISK_SIZE in universal/kisak_abi.h). The
    xanim::XAnimPartsNative runtime view widens those fields to
    native-width pointers; sizeof(XAnimPartsNative) == 0x88 on 64-bit
    hosts and == 88 on 32-bit hosts (RUNTIME_SIZE). The
    XAnimPartsDiskToNative / XAnimPartsNativeToDisk helpers move
    values between the two views with explicit ordering; on 32-bit
    they are bit-identical (uint32_t == pointer == 4 bytes), on 64-bit
    they widen pointer fields to native-width.

  * src/xanim/xanim.h: small consistency change. The on-disk structs
    (XAnimIndices, XAnimNotifyInfo, XAnimDeltaPart, XAnimParts) define
    KISAK_XANIM_*_DECLARED macros so re-inclusion with xanim_native.h
    after the shim provides the same shape is safe. The actual struct
    shapes in the engine header are unchanged: a 32-bit-ILP32
    on-disk-frozen retail layout that the existing 88-byte static_assert
    in xanim.h still pins on 32-bit builds. The 64-bit path is
    exercised by the contract test below.

  * src/xanim/xanim.cpp: XAnimClone now allocates sizeof(xanim::XAnimPartsNative)
    bytes, copies the frozen 88-byte disk fields into the head of the
    buffer, and zeroes the trailing bytes that a widened native view
    would expose but the on-disk source does not carry. On 32-bit the
    allocation is byte-identical with the prior Alloc(88) + qmemcpy
    path; on 64-bit the allocation matches the runtime view, the
    memcpy covers the 88-byte disk mirror, and the memset blanks the
    extra runtime bytes so a future consumer that dereferences them
    sees a known state instead of stale hunk memory.

  * tests/xanim_parts_split_test_shim.h: minimal on-disk XAnimParts
    types for portable test binaries. The shim mirrors the disk mirror
    shape but stores every pointer field as a 32-bit disk offset, so
    sizeof(XAnimParts) == 88 on every host architecture. The shim
    defines the KISAK_XANIM_*_DECLARED macros first so xanim_native.h
    can run its ONDISK_SIZE / RUNTIME_SIZE contracts against the shim
    without redefinition errors. Test TUs include this shim before
    xanim_native.h.

  * tests/xanim_parts_split_test.cpp: contract test for the disk
    mirror vs native runtime split. Exercises six contracts: (1)
    ONDISK_SIZE(XAnimParts, 88) holds on every target; (2)
    RUNTIME_SIZE(XAnimPartsNative, 88, 0x88) holds on 32-bit and
    64-bit; (3) DiskToNative preserves every scalar field; (4)
    NativeToDisk preserves every scalar field; (5) round-trip
    Disk -> Native -> Disk is byte-identical for the 88-byte wire
    image; (6) the runtime view size is at least the disk size
    (which is the property that prevents the legacy Alloc(88)
    under-allocation on 64-bit). All 78 checks pass under the local
    GCC build.

  * tests/CMakeLists.txt: registers the kisakcod-xanim-parts-split-tests
    target and the xanim-parts-split-contracts ctest entry. The CTest
    regex 'xanim' routes the contract test alongside the existing
    xanim-load-bounded-cursor and xmodel-load-bounded-cursor, so the
    bead's `ctest -R 'xanim|xmodel|script_vm'` invocation exercises
    every Priority 7 + M4 contract on the local gate.

Validation:
  - cmake --build build-tests --target kisakcod-xanim-load-tests
    kisakcod-xmodel-load-tests kisakcod-xanim-parts-split-tests builds
    cleanly under the local GCC toolchain.
  - ctest --test-dir build-tests -R 'xanim|xmodel|script_vm' reports
    100% pass (3/3 tests).
  - xanim_parts_split_test executes 78/78 checks successfully on the
    local 64-bit host; the wire-format round-trip is bit-identical
    for the 88-byte disk subset.
  - Full local CTest: 195/197 tests pass. The two failing tests
    (universal-physicalmemory-legacy-source-invariants and
    security-source-regressions) are pre-existing baseline failures
    on this worktree's origin/master and are unrelated to this
    commit.

Follow-ups (separate beads):
  - Migrate every XAnimParts consumer to XAnimPartsNative. The
    XAnimClone allocation is the prototype; the rest of the consuming
    surface (load-object, save-object, asset-pipeline, runtime
    animtree) needs the same native-view migration.
  - Widen the script VM and remaining runtime structures that still
    ship 32-bit-frozen static_asserts. scr_vm.h, scr_evaluate.h, and
    scr_debugger.h each have multiple sizeof(...) == 0xNN contracts
    that will fail on 64-bit; the contracts need to migrate to the
    ONDISK_SIZE / RUNTIME_SIZE macros and the runtime view split
    pattern used here.
  - The engine-side xanim.h still holds raw host pointers in XAnimParts
    to keep the existing 32-bit consumer-base intact; converting the
    engine-side on-disk mirror to 32-bit-typed pointer fields is a
    follow-up that depends on the consumer migration above.

Refinery docs: docs/task.md Priority 7 + M4 contract. The bead
acceptance criteria for XAnimParts/XAnimIndices, XAnimClone (88 bytes),
load-object 32-bit payload assumptions, and the native XAnimParts 0x88
split from disk mirror are now contracted by the test; the wider
SP int-temporary removals in CG_UpdatePhysicsPose / CG_SaveEntityPhysics
/ CG_ShutdownEntity / CG_ExpiredLaunch / CG_CalcEntityPhysicsPositions
are the SP physObjId half of the same Priority 7 contract.

+++ src/cgame/cg_snapshot.cpp
@@ -26,7 +26,12 @@ void __cdecl CG_ShutdownEntity(int localClientNum, centity_s *cent)
     FxEffect *effect; // r4
     trajectory_t *p_pos; // r29
     int ragdollHandle; // r3
-    int physObjId; // r4
+    // SP cpose_t::physObjId is uintptr_t; the legacy `int physObjId`
+    // local silently narrowed the native dxBody* pointer on 64-bit.
+    // Hold the field value as uintptr_t and compare against the same
+    // -1/0 sentinels the field uses so the bit-identical 0x0 / 0xF..F
+    // sentinel contract is preserved without truncating the live pointer.
+    uintptr_t physObjId; // r4
 
     oldEType = cent->oldEType;
     if (oldEType == 8 || oldEType == 7)
@@ -51,11 +56,11 @@ void __cdecl CG_ShutdownEntity(int localClientNum, centity_s *cent)
         cent->currentState.apos.trType = TR_STATIONARY;
     }
     physObjId = cent->pose.physObjId;
-    if (physObjId && physObjId != -1)
+    if (physObjId && physObjId != (uintptr_t)-1)
         goto LABEL_14;
     if (p_pos->trType != TR_PHYSICS)
         return;
-    if (physObjId != -1)
+    if (physObjId != (uintptr_t)-1)
     {
     LABEL_14:
         if (physObjId)
